claims-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
email-validator==2.1.0.post1
openai==1.3.0
python-magic==0.4.27
aiofiles==23.2.1
moviepy==1.0.3 